backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
python-dotenv==1.2.1
pydantic==2.12.5
email-validator==2.3.0
bcrypt==4.0.1
PyJWT==2.10.1
razorpay==2.0.0
resend==2.21.0
python-multipart==0.0.20
passlib==1.7.4
certifi==2024.8.30
